add test for MkCkeck
