+++ src/main/java/com/jcabi/github/Check.java
@@ -30,6 +30,8 @@
 package com.jcabi.github;
 
 import java.io.IOException;
+import java.util.Arrays;
+import java.util.Locale;
 
 /**
  * Github check.
@@ -48,4 +50,183 @@ public interface Check {
      */
     boolean successful() throws IOException;
 
+    /**
+     * Check status.
+     * You can read more about it
+     * @checkstyle LineLengthCheck (1 lines)
+     * <a href="https://docs.github.com/en/rest/checks/runs?apiVersion=2022-11-28#list-check-runs-for-a-git-reference"> here </a>
+     */
+    enum Status {
+        /**
+         * Queued.
+         */
+        QUEUED("queued"),
+
+        /**
+         * In progress.
+         */
+        IN_PROGRESS("in_progress"),
+
+        /**
+         * Completed.
+         */
+        COMPLETED("completed");
+
+        /**
+         * Status.
+         */
+        private final String status;
+
+        /**
+         * Ctor.
+         * @param stat Status.
+         */
+        Status(final String stat) {
+            this.status = stat;
+        }
+
+        /**
+         * Status.
+         * @return Status.
+         */
+        public String value() {
+            return this.status;
+        }
+
+        /**
+         * Get status from string.
+         * @param value String value.
+         * @return Status.
+         */
+        public static Status fromString(final String value) {
+            return Arrays.stream(Status.values())
+                .filter(stat -> stat.same(value))
+                .findFirst()
+                .orElseThrow(
+                    () -> new IllegalArgumentException(
+                        String.format("Invalid value %s for status", value)
+                    )
+                );
+        }
+
+        /**
+         * Check if check is finished.
+         * @return True if check is finished.
+         */
+        boolean finished() {
+            return this == Status.COMPLETED;
+        }
+
+        /**
+         * Check if status is the same as value.
+         * @param value Value.
+         * @return True if status is the same as value.
+         */
+        boolean same(final String value) {
+            return this.status.equals(value.toLowerCase(Locale.ROOT));
+        }
+    }
+
+    /**
+     * Check conclusion.
+     * You can read more about it
+     * @checkstyle LineLengthCheck (1 lines)
+     * <a href="https://docs.github.com/en/rest/checks/runs?apiVersion=2022-11-28#list-check-runs-for-a-git-reference"> here </a>
+     */
+    enum Conclusion {
+
+        /**
+         * Action required.
+         */
+        ACTION_REQUIRED("action_required"),
+
+        /**
+         * Cancelled.
+         */
+        CANCELLED("cancelled"),
+
+        /**
+         * Failure.
+         */
+        FAILURE("failure"),
+
+        /**
+         * Neutral.
+         */
+        NEUTRAL("neutral"),
+
+        /**
+         * Success.
+         */
+        SUCCESS("success"),
+
+        /**
+         * Skipped.
+         */
+        SKIPPED("skipped"),
+
+        /**
+         * Stale.
+         */
+        STALE("stale"),
+
+        /**
+         * Timed out.
+         */
+        TIMED_OUT("timed_out");
+
+        /**
+         * Conclusion.
+         */
+        private final String conclusion;
+
+        /**
+         * Ctor.
+         * @param con Conclusion.
+         */
+        Conclusion(final String con) {
+            this.conclusion = con;
+        }
+
+        /**
+         * Get conclusion from string.
+         * @param value String value.
+         * @return Conclusion.
+         */
+        public static Conclusion fromString(final String value) {
+            return Arrays.stream(Conclusion.values())
+                .filter(stat -> stat.same(value))
+                .findFirst()
+                .orElseThrow(
+                    () -> new IllegalArgumentException(
+                        String.format("Invalid value %s for conclusion", value)
+                    )
+                );
+        }
+
+        /**
+         * Conclusion.
+         * @return Conclusion.
+         */
+        public String value() {
+            return this.conclusion;
+        }
+
+        /**
+         * Check if check is successful.
+         * @return True if check is successful.
+         */
+        boolean successful() {
+            return this == Conclusion.SUCCESS;
+        }
+
+        /**
+         * Check if conclusion is the same as value.
+         * @param value Value to compare.
+         * @return True if conclusion is the same as value.
+         */
+        boolean same(final String value) {
+            return this.conclusion.equals(value.toLowerCase(Locale.ROOT));
+        }
+    }
 }
